+++ CHANGES.md
@@ -24,6 +24,7 @@ Fixes:
 - fix `okla` improve cleanup_content method to use CSS selectors and return bytes #1692
 - Fix `fla` and `fladistctapp` scrapers by scraping new API endpoint #1690
 - fix `uscgcoca` and `asbca` headers #1664
+- `okla` `cleanup_content` now deletes HTML comments that were causing duplicates #1697
 
 ## Current